requirements.txt
streamlit==1.28.0
tensorflow==2.15.0
pillow==10.0.0
numpy==1.24.3
